Include all expenses when retrieving an entry
This changeset updates the GET entry handler to return all
associated expenses along with it. It also adds a method to
the client to retrieve an entry, and an end-to-end test for
the new functionality.

The inclusion of expenses is compatible with the jsonapi
standard: https://jsonapi.org/format/#fetching-includes.
The "include" keyword will be added subsequently.