Add flag for building docker image in e2e script
Whenever the code changes, the casheer image needs rebuilding.
Before, I would add the --build argument in the script manually
if I wanted to run the end-to-end tests just after making some
changes. This adds an option to pass in a --build argument to
the end-to-end testing script to inform whether or not the
image needs rebuilding.