Add include query param for retrieving an entry
This changeset allows for specifying whether or not the expenses
should be returned when making a request to retrieve a single entry.
This can be done by passing in the "include" query parameter, as per
the jsonapi specification: https://jsonapi.org/format/#fetching-includes

Only the GET entry endpoint supports this behavior for now. It doesn't
really make sense to extend the semantics of the other endpoints to
allow the include query parameter at the moment.

The client had been updated to support this change, and an end-to-end
test has been added to test both behaviours.

Closes #7