Add list entries for period to client
Currently the client only supports listing all entries. However,
the expense planning will be done on a month-to-month basis. This
means that from a user perspective, you're typically only interested
in the entries of a given period.

This change adds a method to the client that allows filtering the
entries based off period, alongside a corresponding end-to-end test.