Remove configer and original config dependencies
Previously, this was using quite a complex config solution with Viper,
which is a relatively stale package with a lot of old dependencies.
Nowadays, in a containerized environment, only using environment variables
is pretty much production-ready for configs, let alone for this personal
project. This change removes all the old configuration solution and its
dependencies, in favour of simply reading some env vars instead.