Add domain models
This is the first step towards the transition to DDD, and does the
following:

- port over the definition of the old gorm models (Entry, Expense, Debt)
to storage agnostic aggregates
- define helper value objects for currency and month
- remove validators by providing public methods which always preserve
the consistency of the models

Given the application is written in Go and there is no rich support for
classes, the consistency rules are not "enforced" in the structs, thus
the properties are public. This should be good enough.